web-services/requirements.txt
fastapi[all]
pydantic 
requests
python-dotenv
authlib
